+++ tests/Arbitrary.hs
@@ -7,9 +7,7 @@ import Test.QuickCheck
 
 import qualified Data.ByteString as B
 import qualified Data.ByteString.Lazy as L
-#if MIN_VERSION_bytestring(0,10,4)
 import qualified Data.ByteString.Short as S
-#endif
 
 instance Arbitrary L.ByteString where
   arbitrary = fmap L.fromChunks arbitrary
@@ -17,7 +15,5 @@ instance Arbitrary L.ByteString where
 instance Arbitrary B.ByteString where
   arbitrary = B.pack `fmap` arbitrary
 
-#if MIN_VERSION_bytestring(0,10,4)
 instance Arbitrary S.ShortByteString where
   arbitrary = S.toShort `fmap` arbitrary
-#endif